+++ tsdown.config.ts
@@ -1,4 +1,3 @@
-import { defaultIgnore, generateDTS } from '@jakeboone02/generate-dts';
 import { writeFile } from 'fs/promises';
 import type { UserConfig } from 'tsdown';
 import { defineConfig } from 'tsdown';
@@ -8,7 +7,6 @@ const config: ReturnType<typeof defineConfig> = defineConfig(options => {
     entry: {
       'numeric-quantity': 'src/index.ts',
     },
-    dts: false,
     platform: 'neutral',
     sourcemap: true,
     ...options,
@@ -25,13 +23,6 @@ const config: ReturnType<typeof defineConfig> = defineConfig(options => {
       ...commonOptions,
       clean: true,
       format: 'esm',
-      onSuccess: () =>
-        generateDTS({
-          ignore: filePath =>
-            defaultIgnore(filePath) ||
-            filePath.endsWith('Tests.ts') ||
-            filePath.endsWith('/dev.ts'),
-        }),
     },
     // ESM, Webpack 4 support. Target ES2017 syntax to compile away optional chaining and spreads
     {
@@ -74,16 +65,22 @@ const config: ReturnType<typeof defineConfig> = defineConfig(options => {
       outDir: './dist/cjs/',
       onSuccess: async () => {
         // Write the CJS index file
-        await writeFile(
-          'dist/cjs/index.js',
-          `'use strict';
+        await Promise.all([
+          writeFile(
+            'dist/cjs/index.js',
+            `'use strict';
 if (process.env.NODE_ENV === 'production') {
   module.exports = require('./numeric-quantity.cjs.production.js');
 } else {
   module.exports = require('./numeric-quantity.cjs.development.js');
 }
 `
-        );
+          ),
+          writeFile(
+            'dist/cjs/index.d.ts',
+            `export * from './numeric-quantity.cjs.development.js';`
+          ),
+        ]);
       },
     },
     // UMD (ish)